Revert "feat: use single-file components"
This reverts commit 1e08f68840937bb9152827b290c76873de5ba927.
