+++ apps/this-is-learning/src/app/app.component.spec.ts
@@ -1,12 +1,13 @@
 import { TestBed } from '@angular/core/testing';
 import { RouterTestingModule } from '@angular/router/testing';
 
-import { AppComponent, AppScam } from './app.component';
+import { AppComponent } from './app.component';
 
 describe('AppComponent', () => {
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      imports: [AppScam, RouterTestingModule],
+      declarations: [AppComponent],
+      imports: [RouterTestingModule],
     }).compileComponents();
   });
