+++ apps/this-is-learning/src/app/app.module.ts
@@ -1,12 +1,13 @@
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
+import { ScullyLibModule } from '@scullyio/ng-lib';
 
 import { AppRoutingModule } from './app-routing.module';
-import { AppComponent, AppScam } from './app.component';
-import { ScullyLibModule } from '@scullyio/ng-lib';
+import { AppComponent } from './app.component';
 
 @NgModule({
   bootstrap: [AppComponent],
-  imports: [BrowserModule, AppScam, AppRoutingModule, ScullyLibModule],
+  declarations: [AppComponent],
+  imports: [BrowserModule, AppRoutingModule, ScullyLibModule],
 })
 export class AppModule {}
